Predict Method for ULDA Model

This function predicts the class labels or class probabilities for new data using a fitted ULDA model. The prediction can return either the most likely class ("response") or the posterior probabilities for each class ("prob").

Usage

# S3 method for class 'ULDA'
predict(object, newdata, type = c("response", "prob"), ...)

Arguments

object

A fitted ULDA model object.

newdata

A data frame containing the new predictor variables for which predictions are to be made.

type

A character string specifying the type of prediction to return. "response" returns the predicted class labels, while "prob" returns the posterior probabilities for each class. Default is "response".

...

Additional arguments.

Value

If type = "response", the function returns a vector of predicted class labels. If type = "prob", it returns a matrix of posterior probabilities, where each row corresponds to a sample and each column to a class.

Examples

fit <- folda(datX = iris[, -5], response = iris[, 5], subsetMethod = "all")

# Predict class labels
predictions <- predict(fit, iris, type = "response")

# Predict class probabilities
prob_predictions <- predict(fit, iris, type = "prob")